chess/src/test/java/ru/job4j/chess/LogicTest.java
package ru.job4j.chess;

import org.junit.Test;
import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertThat;

import ru.job4j.chess.firuges.Cell;
import ru.job4j.chess.firuges.black.BishopBlack;
import ru.job4j.chess.firuges.black.PawnBlack;

public class LogicTest {

    Logic logic = new Logic();

    /**
     * Если не добавлять фигуры в массив класса Logic,
     * должен сработать класс исключение FigureNotFoundException
     */
    @Test (expected = FigureNotFoundException.class)
    public void figureNotFound(){
        logic.move(Cell.E7, Cell.E5);
    }

    /**
     * Не очень понятно, что именно попадает в массив
     * Cell[] steps метода move(), попробуем
     * сделать разрешенное движение - черная пешка ходит на одну клетку
     * и запустить дебаг
     *
     * Результат:
     * Cell[] steps возвращает клетку назначения, если ход по правилам и
     * Cell[0] если ход не по правилам.
     * После этого разрешает (или не разрешает перемещение)
     * this.figures[index].copy(dest)
     */
    @Test
    public void pawnBlackMove(){
        logic.add(new PawnBlack(Cell.E7));
        boolean result = logic.move(Cell.E7, Cell.E6);
        assertThat(result, is(true));
    }

    /**
     * метод findBy() можно использовать для проверки
     * осутствия фигур на всем пути фигуры
     * Для этого мы можем добавить исключение
     * OccupiedWayException в метод move()
     * и проверить это
     */
    @Test (expected = OccupiedWayException.class)
    public void pawnCannotMove(){
        logic.add(new PawnBlack(Cell.E7));
        logic.add(new PawnBlack(Cell.E6));
        logic.move(Cell.E7,  Cell.E6);
    }

    /**
     * Проверим 3 исключение ImpossibleMoveException
     * которое должно создаваться, когда пешка ходит
     * неправильно, например на 2 клетки
     */
    @Test (expected = ImpossibleMoveException.class)
    public void pawnImpossibleMove(){
        logic.add(new PawnBlack(Cell.E7));
        logic.move(Cell.E7, Cell.E5);
    }

    /**
     * Check Bishop black
     * проверяем ходы черного слона, в нормальных условиях
     * через Logic
     */
    @Test
    public void bishopBlackMove() {
        logic.add(new BishopBlack(Cell.C8));
        boolean result = logic.move(Cell.C8, Cell.F5);
        assertThat(result, is(true));
    }

    /**
     * Check Bishop black 2
     * проверяем в нормальных условиях как работает way()
     * Такие же значения, как в предыдущем тесте
     * expect - сохраняются значения все кроме первого
     */
    @Test
    public void bishopBlackWay() {
        BishopBlack bishopBlack = new BishopBlack(Cell.C8);
        Cell[] result = bishopBlack.way(Cell.C8, Cell.F5);
        Cell[] expect = {Cell.D7, Cell.E6, Cell.F5};
        assertThat(result, is(expect));
    }
}
